+++ src/ska_tmc_centralnode/central_node.py
@@ -43,6 +43,16 @@ class AbstractCentralNode(TMCBaseDevice):
         default_value=tuple(),
     )
 
+    AssignResourcesInterface = device_property(
+        dtype="str",
+        doc="Interface value of AsignResources schema",
+    )
+
+    ReleaseResourcesInterface = device_property(
+        dtype="str",
+        doc="Interface value of ReleaseResources schema",
+    )
+
     CspMasterLeafNodeFQDN = device_property(dtype="str")
 
     CspMasterFQDN = device_property(dtype="str")