+++ src/ska_tmc_centralnode/manager/component_manager_low.py
@@ -47,6 +47,12 @@ class CNComponentManagerLow(CNComponentManager):
         sleep_time=1,
         skuid_service="",
         command_timeout=30,
+        assignresources_interface: str = (
+            "https://schema.skao.int/ska-low-tmc-assignresources/4.0"
+        ),
+        releaseresources_interface: str = (
+            "https://schema.skao.int/ska-low-tmc-releaseresources/3.0"
+        ),
         *args,
         **kwargs,
     ):
@@ -88,6 +94,8 @@ class CNComponentManagerLow(CNComponentManager):
             sleep_time,
             skuid_service=skuid_service,
             command_timeout=command_timeout,
+            assignresources_interface=assignresources_interface,
+            releaseresources_interface=releaseresources_interface,
             *args,
             **kwargs,
         )